Move docs.cli to scripts
