+++ packages/mermaid/scripts/docs.cli.mts
