+++ packages/mermaid/scripts/docs.mts
