+++ packages/mermaid/scripts/docs.spec.ts
